+++ src/createKeydownFromClick/componentClass.test.ts
@@ -1,4 +1,0 @@
-import { ComponentClass } from '../__fixtures__/ComponentClass'
-import { test } from '../utils/test'
-
-test('createKeydownFromClick', ComponentClass)