+++ src/tests/componentClass.test.ts
@@ -1,0 +1,4 @@
+import { ComponentClass } from './__fixtures__/ComponentClass'
+import { runGeneralTests } from './utils'
+
+runGeneralTests('createKeydownFromClick', ComponentClass)