+++ src/tests/typings.d.ts
@@ -1,8 +1,10 @@
 import { FireFunction } from '@testing-library/react'
 
-import { Modifier, Modifiers } from '../../typings'
+import { Modifier, Modifiers } from '../typings'
 
-export type MockedEvent = { key: KeyboardEvent['key'] }
+export type MockedEvent = {
+  key: KeyboardEvent['key']
+}
 
 export type MockedEventWithModifiers = MockedEvent & Modifiers