+++ src/utils/console.ts
@@ -1,0 +1,2 @@
+export const warn = (message: string): void =>
+  console.warn('[keydown-from-click]: ' + message)